# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "คณะครุศาสตร์ มหาวิทยาลัยราชภัฏราชนครินทร์ (EDU RRU) - RESTful API",
        "description": "ระบบ RESTful API ครอบคลุมการทำงานของระบบสารสนเทศ เว็บไซต์หลัก และระบบบริหารจัดการหลังบ้าน (Backoffice) คณะครุศาสตร์ มหาวิทยาลัยราชภัฏราชนครินทร์\n\n### ฟังก์ชันหลักของระบบ:\n- **Dashboard & Stats:** สรุปจำนวนข่าวสาร บุคลากร สาขาวิชา จำนวนเข้าชม\n- **Posts & SDGs:** ข่าวสาร กิจกรรม บทความ เชื่อมโยงเป้าหมายการพัฒนาที่ยั่งยืน (SDGs 1-17)\n- **Categories:** บริหารจัดการหมวดหมู่ข่าวสารและสี Badge ประจำหมวดหมู่\n- **Personnel & Departments:** ฐานข้อมูลคณาจารย์ บุคลากร และสาขาวิชาประจำคณะครุศาสตร์\n- **Executives:** โครงสร้างผู้บริหารคณะ การจัดหมวดหมู่ และการจัดเรียงลำดับ\n- **Curricula:** หลักสูตรการศึกษา (ปริญญาตรี ค.บ., ประกาศนียบัตรบัณฑิต ป.บัณฑิต, ปริญญาโท ค.ม.)\n- **Regulations & Law:** ข้อบังคับ ระเบียบ ประกาศ คำสั่งมหาวิทยาลัย และหมวดหมู่กฎหมาย\n- **ITA:** ข้อมูลการประเมินคุณธรรมและความโปร่งใส (Integrity and Transparency Assessment)\n- **Carousel:** ภาพสไลด์แบนเนอร์ประชาสัมพันธ์หน้าแรก พร้อมลิงก์ปลายทาง\n- **Philosophy & Committee:** ปรัชญา วิสัยทัศน์ พันธกิจ และคณะกรรมการประจำคณะ\n- **Uploads:** ระบบจัดเก็บไฟล์ภาพและเอกสารแนบ (Storage Disk)",
        "contact": {
            "name": "คณะครุศาสตร์ มหาวิทยาลัยราชภัฏราชนครินทร์",
            "url": "https://edu.rru.ac.th",
            "email": "edu@rru.ac.th"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
...
    "paths": {
...
        "/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "ดูรายละเอียดข่าวสารตาม ID (พร้อมนับยอดเข้าชม +1)",
                "description": "ส่งกลับข้อมูลข่าวสารฉบับเต็ม รวมถึงเนื้อหา HTML, ภาพปก, ภาพแกลเลอรี, ไฟล์แนบ, และรายการ SDGs ที่สอดคล้อง",
                "operationId": "4b1b11f1a637b18ff52426e7a38705d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "รหัสข่าวสาร (Post ID)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "พบข้อมูลข่าวสาร"
                    },
                    "404": {
                        "description": "ไม่พบข่าวสารที่ระบุ"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "แก้ไขข้อมูลข่าวสาร",
                "description": "อัปเดตข้อมูลข่าวสารตาม ID ที่ระบุ",
                "operationId": "e860df2a1150366605f40ebd07bf8385",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "อัปเดตข่าวสารสำเร็จ"
                    },
                    "404": {
                        "description": "ไม่พบข่าวสาร"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "ลบข่าวสาร",
                "description": "ลบข่าวสารออกจากระบบอย่างถาวร",
                "operationId": "5cc62b23119c8525e722545ba9e0984f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ลบข่าวสารสำเร็จ"
                    }
                }
            }
        },
...
    },
...
}